Move base class out of __init__
- this is a bad design pattern and is likely to lead to circular imports
