+++ pooltool/ani/modes/pick_ball.py
@@ -4,8 +4,9 @@
 import numpy as np
 
 import pooltool.ani as ani
+import pooltool.ani.action as action
 import pooltool.constants as c
-from pooltool.ani.modes import Mode, action
+from pooltool.ani.modes.datatypes import Mode
 
 
 class PickBallMode(Mode):
