+++ pooltool/ani/modes/shot.py
@@ -4,8 +4,9 @@ import numpy as np
 
 import pooltool as pt
 import pooltool.ani as ani
+import pooltool.ani.action as action
 import pooltool.ani.utils as autils
-from pooltool.ani.modes import Mode, action
+from pooltool.ani.modes.datatypes import Mode
 
 
 class ShotMode(Mode):
