+++ pooltool/ani/modes/stroke.py
@@ -1,7 +1,8 @@
 #! /usr/bin/env python
 
 import pooltool.ani as ani
-from pooltool.ani.modes import Mode, action
+import pooltool.ani.action as action
+from pooltool.ani.modes.datatypes import Mode
 
 
 class StrokeMode(Mode):
